Fix a division by zero in the GPS interpolation
A lot of gpxpy versions drop sub-second timestamp information which
typically results in duplicate timestamps at the start or end of the
recording (e.g. with the Mapillary app) despite slightly different
latitude or longitude.
So to avoid a division by zero, check that the denominator (the
timestamp difference) is non-zero, and before performing the
division!

Signed-off-by: Francois Gouget <fgouget@free.fr>