Centralize and standardize the --skip_subfolders and --rerun options
Most commands take these two options so this saves a lot of duplication
and groups them with the related --import_path option.

Also capitalize each option's help message.
Put the option on the add_argument() line, the help message on the next
line with a 4 space indentation so short help messages have a chance to
not wrap, and additional options on the following lines so they are all
aligned.

Signed-off-by: Francois Gouget <fgouget@free.fr>