Centralize the download command line option declarations
This way they are declared in the module that uses them and can be
reused without code duplication.

Also capitalize each option's help message.
Put the option on the add_argument() line, the help message on the next
line with a 4 space indentation so short help messages have a chance to
not wrap, and additional options on the following lines so they are all
aligned.

Signed-off-by: Francois Gouget <fgouget@free.fr>